+++ packages/core-data/src/selectors.ts
@@ -344,7 +344,7 @@ export const getEntityRecord = createSelector(
 				const field = fields[ f ].split( '.' );
 				let value = item;
 				field.forEach( ( fieldName ) => {
-					value = value[ fieldName ];
+					value = value?.[ fieldName ];
 				} );
 				setNestedValue( filteredItem, field, value );
 			}